blog-1.md
A union type is a type that contains a number of values ​​in that type, which are called stream literals, and if one of these conditions is true, the task will be completed, and an intersector is an operator that joins two types. If you want to give all values ​​then inter sector will work otherwise it will show error in inter sector all conditions must be fulfilled but in union if all conditions are not fulfilled.